classes: Changed shader check level to 45: compute shader support.
During the CheckCompatibility method it checks for several things the
target must support, which one of them is the graphics shader level. The
comment specifies "For compute shader support", which is level 45.
Compatibility is checked against level 50, which is DX11. According to
a reddit response (and lack of target support information) macos is not
to be excluded. DX is obviously not supported on macos.